refactor : modal에서 invitation 정보를 호출하도록 변경
- mypage에서 정보를 넘겨주는 것이 아닌 modal에서 정보를 호출하도록 변경함
- date의 second의 경우 소수점을 출력하여 반올림 함수를 적용해줌
