+++ client/src/components/molecules/InvitationCard/InvitationCard.stories.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import InvitationCard from './InvitationCard';
 
 interface Props {
+  id: number;
   master: string;
   name: string;
   time: string;
@@ -11,6 +12,7 @@ export default {
   title: 'molecules/InvitationCard',
   component: InvitationCard,
   argTypes: {
+    id: { control: 'number' },
     master: { control: 'text' },
     name: { control: 'text' },
     time: { control: 'text' },
@@ -18,9 +20,11 @@ export default {
 };
 
 export const invitationCard = ({ ...args }: Props): JSX.Element => {
-  const callback = () => {
+  const callback = (id: number, isAccept: boolean) => {
+    console.log(id, isAccept);
     return true;
   };
+
   return <InvitationCard {...args} callback={callback} />;
 };
 
@@ -29,6 +33,7 @@ invitationCard.story = {
 };
 
 invitationCard.args = {
+  id: 1,
   master: '김도연',
   name: '부스트 동아리',
   time: '2020-11-10 02:12:00',
