+++ client/src/views/MyPage.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import CenterContent from '@molecules/CenterContent';
 import BeeBackground from '@organisms/BeeBackground';
 import TopNavBar from '@organisms/TopNavBar';
@@ -9,25 +9,11 @@ import CreditCardEditModal from '@organisms/CreditCardEditModal';
 import UserInfoSettingModal from '@organisms/UserInfoSettingModal';
 import { useSelector } from 'react-redux';
 import { RootState } from '@reducers/rootReducer';
-import { Invitation } from '@interfaces/accountbook';
-import { getAxiosData } from '@utils/axios';
-import * as API from '@utils/api';
 
 const MyPage: React.FC = () => {
   const modalView = useSelector((state: RootState) => state.modal.view);
   const userName = useSelector((state: RootState) => state.user.name);
   const userImage = useSelector((state: RootState) => state.user.image);
-
-  const [invitations, setInvitations] = useState<Invitation[]>([]);
-
-  const getInvitation = async () => {
-    const result = await getAxiosData(API.GET_INVITATION);
-    setInvitations(result.data);
-  };
-
-  useEffect(() => {
-    getInvitation();
-  }, []);
 
   return (
     <>
@@ -37,7 +23,7 @@ const MyPage: React.FC = () => {
         <MyPageUserMenu name={userName} profile={userImage} />
         <MyPageMenu />
       </CenterContent>
-      {modalView === 'AccountBookAccept' && <InvitationManagementModal data={invitations} />}
+      {modalView === 'AccountBookAccept' && <InvitationManagementModal />}
       {modalView === 'CreditCardEdit' && <CreditCardEditModal />}
       {modalView === 'UserInfoSetting' && <UserInfoSettingModal />}
     </>
